ESP/main: Separate CN_WIRED driver
Restructure the code in a manner, similar to ESP8266 port. Introduce a
replace-able CN_WIRED driver implementation in a separate file.
