+++ ESP/main/CMakeLists.txt
@@ -1,3 +1,3 @@
-set (COMPONENT_SRCS "Faikin.c" "bleenv.c" "../settings.c")
+set (COMPONENT_SRCS "cn_wired_driver.c" "Faikin.c" "bleenv.c" "../settings.c")
 set (COMPONENT_REQUIRES "ESP32-RevK")
 register_component ()
