+++ ESP/main/cn_wired.h
@@ -1,0 +1,7 @@
+#ifndef CN_WIRED_H
+#define CN_WIRED_H
+
+// Data packet length
+#define CN_WIRED_LEN 8
+
+#endif
